requirements.txt
python-dotenv
httpx
requests
jwt